RSL functional

+++ projet/stsp2graph.jl
@@ -23,7 +23,7 @@ function stsp2graph(filename)
 
     """ Initialisation du graphe avec le premier noeud et la première arrête convertis"""
 
-    g=Graph("MyGraph",[Node(string(1),dict_nodes[1])],[Edge(Node(string(list_edges[1][1]+1),dict_nodes[list_edges[1][1]+1]),Node(string(list_edges[1][2]+1),dict_nodes[list_edges[1][2]+1]),parse(Int64,list_edges[1][3]))])
+    g=Graph("MyGraph",[Node(string(1),dict_nodes[1])],[Edge(Node(string(list_edges[1][1]+1),dict_nodes[list_edges[1][1]+1]),Node(string(list_edges[1][2]+1),dict_nodes[list_edges[1][2]+1]),parse(Float64,list_edges[1][3]))])
 
     """ Ajout des noeuds"""
     i=2
@@ -36,7 +36,7 @@ function stsp2graph(filename)
     for i in 2:length(list_edges)
         firstnode=Node(string(list_edges[i][1]+1),dict_nodes[list_edges[i][1]+1])
         secondnode=Node(string(list_edges[i][2]+1),dict_nodes[list_edges[i][2]+1])
-        edgeweight=parse(Int64,list_edges[i][3])
+        edgeweight=parse(Float64,list_edges[i][3])
         add_edge!(g,Edge(firstnode,secondnode,edgeweight))
     end
